Refactor... Moved XDSUser model to users app Moved permission checker class to users app Moved password validators to users app Moved XDSConfiguration to configurations app Moved XDSUIConfiguration to configurations app Various DB migrations Moved tests to relevant apps Updated XDS-API views to be class based
