+++ app/configurations/tests/test_views_integration.py
@@ -1,11 +1,13 @@
 import json
+from django.core.exceptions import ValidationError
 
-from django.test import tag
+from django.test import tag, TestCase
 from django.urls import reverse
 from rest_framework import status
 from rest_framework.test import APITestCase
 
-from core.models import SearchSortOption, XDSConfiguration, XDSUIConfiguration
+from core.models import SearchSortOption
+from configurations.models import XDSConfiguration, XDSUIConfiguration
 
 
 @tag('integration')
@@ -14,7 +16,7 @@ class ViewTests(APITestCase):
     def test_xds_ui_config_view(self):
         """Test that making a GET request to the api gives us a JSON of the
             stored XDSUIConfiguration model AND its sort options"""
-        url = reverse('xds_api:xds-ui-configuration')
+        url = reverse('configurations:xds-ui-configuration')
         xds_config = XDSConfiguration(target_xis_metadata_api="test")
         xds_ui_cfg = XDSUIConfiguration(search_results_per_page=10,
                                         xds_configuration=xds_config)
@@ -41,3 +43,33 @@ class ViewTests(APITestCase):
                          xds_ui_cfg.search_results_per_page)
         self.assertEqual(len(response_dict['search_sort_options']), 2)
         self.assertEqual(response.status_code, status.HTTP_200_OK)
+
+
+@tag('integration')
+class ModelTests(TestCase):
+
+    def test_create_two_xds_configuration(self):
+        """Test that trying to create more than one XDS Configuration objects
+            throws ValidationError """
+        with self.assertRaises(ValidationError):
+            xdsConfig = XDSConfiguration(target_xis_metadata_api="test")
+            xdsConfig2 = XDSConfiguration(target_xis_metadata_api="test2")
+            xdsConfig.save()
+            xdsConfig2.save()
+
+    def test_xds_ui_config_rpp_validator(self):
+        """Test that creating an XSD config object with a value lower than the
+            min value triggers a validation error"""
+        xds_config = XDSConfiguration(target_xis_metadata_api="test")
+        xds_ui_config = XDSUIConfiguration(search_results_per_page=-1,
+                                           xds_configuration=xds_config)
+        self.assertRaises(ValidationError, xds_ui_config.full_clean)
+
+    def test_xds_config_save_success(self):
+        """Test that creating an XDS config object with correct values passes
+            validation"""
+        xdsConfig = XDSConfiguration(target_xis_metadata_api="test")
+        xdsConfig.save()
+        retrievedObj = XDSConfiguration.objects.first()
+
+        self.assertEqual(retrievedObj.target_xis_metadata_api, "test")
