+++ app/es_api/utils/queries.py
@@ -5,8 +5,8 @@ import os
 from elasticsearch_dsl import A, Document, Q, Search, connections
 from elasticsearch_dsl.query import MoreLikeThis
 
-from core.models import (CourseSpotlight, SearchFilter, SearchSortOption,
-                         XDSConfiguration)
+from core.models import (CourseSpotlight, SearchFilter, SearchSortOption)
+from configurations.models import XDSConfiguration
 
 connections.create_connection(alias='default',
                               hosts=[os.environ.get('ES_HOST'), ], timeout=60)
