+++ app/openlxp_xds_project/urls.py
@@ -22,6 +22,8 @@ from django.urls import include, path
 urlpatterns = [
     path('admin/', admin.site.urls),
     path('api/', include('xds_api.urls')),
+    path('api/', include('users.urls')),
+    path('api/', include('configurations.urls')),
     path('es-api/', include('es_api.urls')),
     # url('', include('social_django.urls', namespace='social')),
     path('api-auth/', include('rest_framework.urls',
