+++ app/xds_api/tests/test_xds_utils_unit.py
@@ -3,7 +3,8 @@ from unittest.mock import patch
 
 from django.test import TestCase, tag
 
-from core.models import CourseSpotlight, Experience, XDSConfiguration
+from core.models import CourseSpotlight, Experience
+from configurations.models import XDSConfiguration
 from xds_api.utils.xds_utils import (get_spotlight_courses_api_url,
                                      metadata_to_target, save_experiences)
