+++ app/xds_api/urls.py
@@ -7,43 +7,35 @@ router = DefaultRouter()
 
 app_name = 'xds_api'
 urlpatterns = [
-    path('configuration/', views.XDSConfigurationView.as_view(),
-         name='xds-configuration'),
-    path('ui-configuration/', views.XDSUIConfigurationView.as_view(),
-         name='xds-ui-configuration'),
-    path('spotlight-courses', views.get_spotlight_courses,
+    path('spotlight-courses', views.GetSpotlightCoursesView.as_view(),
          name='spotlight-courses'),
-    path('auth/register', views.RegisterView.as_view(), name='register'),
-    path('auth/login', views.LoginView.as_view(), name='login'),
-    path('auth/logout', views.logout_view, name='logout'),
-    path('auth/validate', views.is_logged_in, name='validate'),
-    path('experiences/<str:exp_hash>/', views.get_experiences,
+    path('experiences/<str:exp_hash>/', views.GetExperiencesView.as_view(),
          name='get_courses'),
-    path('interest-lists/', views.interest_lists,
+    path('interest-lists/', views.InterestListsView.as_view(),
          name='interest-lists'),
-    path('interest-lists/<int:list_id>', views.interest_list,
+    path('interest-lists/<int:list_id>', views.InterestListView.as_view(),
          name='interest-list'),
     path('experiences/<str:exp_hash>/interest-lists',
-         views.add_course_to_lists,
+         views.AddCourseToListsView.as_view(),
          name='add_course_to_lists'),
     path('interest-lists/owned',
-         views.interest_lists_owned,
+         views.InterestListsOwnedView.as_view(),
          name='owned-lists'),
     path('interest-lists/subscriptions',
-         views.interest_lists_subscriptions,
+         views.InterestListsSubscriptionsView.as_view(),
          name='interest-list-subscriptions'),
     path('interest-lists/<int:list_id>/subscribe',
-         views.interest_list_subscribe,
+         views.InterestListSubscribeView.as_view(),
          name='interest-list-subscribe'),
     path('interest-lists/<int:list_id>/unsubscribe',
-         views.interest_list_unsubscribe,
+         views.InterestListUnsubscribeView.as_view(),
          name='interest-list-unsubscribe'),
-    path('saved-filters/<int:filter_id>', views.saved_filter,
+    path('saved-filters/<int:filter_id>', views.SavedFilterView.as_view(),
          name='saved-filter'),
     path('saved-filters/owned',
-         views.saved_filters_owned,
+         views.SavedFiltersOwnedView.as_view(),
          name='owned-filters'),
     path('saved-filters',
-         views.saved_filters,
+         views.SavedFiltersView.as_view(),
          name='saved-filters'),
 ]
